pyproject.toml
[tool.poetry]
name = "k8s-resource-bot"
version = "0.0.1"
description = "A python based applications that analyses compute resources configured in a kubernetes namespace with values configured in YAML files."
authors = ["jaydee94 <janherber@gmx.de>"]
license = "Apache-2.0 License"

[tool.poetry.dependencies]
python = "^3.8"
json-logging = "^1.3"
click = "^8"
pyaml = "*"
pydantic = "^1.9.0"
kubernetes = "^23.3.0"
python-json-logger = "^2.0.2"

[tool.poetry.dev-dependencies]
black = "*"
flake8 = "*"
flake8-awesome = "*"
isort = "*"
pytest = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
k8s-resource-bot = "k8s_resource_bot:cmd"
